[#34] refactor: 패키지 리팩토링
