+++ BE/src/main/java/com/codesquad/issue/domain/issue/response/IssueResponse.java
@@ -1,5 +1,6 @@
-package com.codesquad.issue.domain.issue;
+package com.codesquad.issue.domain.issue.response;
 
+import com.codesquad.issue.domain.issue.Issue;
 import lombok.Getter;
 import lombok.ToString;
