+++ BE/src/main/java/com/codesquad/issue/service/IssueService.java
@@ -3,9 +3,9 @@ package com.codesquad.issue.service;
 import com.codesquad.issue.domain.account.Account;
 import com.codesquad.issue.domain.account.AccountResponse;
 import com.codesquad.issue.domain.issue.Issue;
-import com.codesquad.issue.domain.issue.IssueDetailResponse;
+import com.codesquad.issue.domain.issue.response.IssueDetailResponse;
 import com.codesquad.issue.domain.issue.IssueRepository;
-import com.codesquad.issue.domain.issue.IssueResponse;
+import com.codesquad.issue.domain.issue.response.IssueResponse;
 import com.codesquad.issue.domain.issue.request.IssueCreateRequest;
 import com.codesquad.issue.global.error.exception.IssueNotFoundException;
 import lombok.RequiredArgsConstructor;
